Phone-Book/phonebook_project/requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.2
django-coverage-plugin==1.3.1
Sphinx==1.4.8
django-extensions==1.7.4
Werkzeug==0.11.11
django-test-plus==1.0.16
factory_boy==2.7.0

django-debug-toolbar==1.6

# improved REPL
ipdb==0.10.1

pytest-django==3.0.0
pytest-sugar==0.7.1

# For phonenumbers
django-phonenumber-field==1.1.0
